+++ CHANGELOG.md
@@ -29,6 +29,10 @@
   - Added URI support: `group://<name>`
   - Added object support: `group: '<name>'`
   - This is similar to the `@files`, `@globs`, etc, token functions.
+- Added URI and object support for task output types, similar to inputs.
+  - `file://<path>` and `file: '<path>'`
+  - `glob://<pattern>` and `glob: '<pattern>'`
+  - Added `optional` support to file outputs.
 - Added new values to the `runInCI` task option:
   - `only` - Only run the task in CI, and not locally, when affected.
   - `skip` - Skip running in CI but run locally and allow task relationships to be valid.